Added patch map generation, added working spawn system (at least
clouds spawn), and a bunch of other stuff needed to make that work,
started on microbe organelle placing

+++ src/microbe_stage/ISpawner.cs
@@ -32,8 +32,8 @@ public abstract class ISpawner
 
     public abstract List<ISpawned> Spawn(Node worldNode, Vector3 location);
 
-    public void SetFrequencyFromDensity(int spawnDensity)
+    public void SetFrequencyFromDensity(float spawnDensity)
     {
-        SpawnFrequency = spawnDensity * SpawnRadiusSqr * 4;
+        SpawnFrequency = (int)(spawnDensity * SpawnRadiusSqr * 4);
     }
 }
